Interrupt on error when building docs locally. Added switch for faster inner loop.

+++ docs/build-dev.ps1
@@ -1,17 +1,37 @@
-# This script assumes that you have already installed docfx and httpserver.
+#Requires -Version 7.0
+
+# This script builds the documentation website, starts a web server and opens the site in your browser. Intended for local development.
+# It is assumed that you have already installed docfx and httpserver.
 # If that's not the case, run the next commands:
 #   choco install docfx -y
 #   npm install -g httpserver
 
-Remove-Item _site -Recurse -ErrorAction Ignore
+param(
+    # Specify -NoBuild to skip code build and examples generation. This runs faster, so handy when only editing Markdown files.
+    [switch] $NoBuild=$False
+)
 
-dotnet build .. --configuration Release
-Invoke-Expression ./generate-examples.ps1
+function VerifySuccessExitCode {
+    if ($LastExitCode -ne 0) {
+        throw "Command failed with exit code $LastExitCode."
+    }
+}
+
+if (-Not $NoBuild -Or -Not (Test-Path -Path _site)) {
+    Remove-Item _site -Recurse -ErrorAction Ignore
+
+    dotnet build .. --configuration Release
+    VerifySuccessExitCode
+
+    Invoke-Expression ./generate-examples.ps1
+}
 
 docfx ./docfx.json
-Copy-Item home/*.html _site/
-Copy-Item home/*.ico _site/
-Copy-Item -Recurse home/assets/* _site/styles/
+VerifySuccessExitCode
+
+Copy-Item -Force home/*.html _site/
+Copy-Item -Force home/*.ico _site/
+Copy-Item -Force -Recurse home/assets/* _site/styles/
 
 cd _site
 $webServerJob = httpserver &
